requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# TUI dependencies
rich>=13.0.0
textual>=0.41.0



# Web dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Additional features
cryptography>=41.0.0
schedule>=1.2.0
click>=8.1.0
pydantic>=2.4.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
croniter>=2.0.0

# Binary building
pyinstaller>=6.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0